Use tying the knot
When we are working with lazy language like haskell,
it's a common case when we are using special techniques
like tying the knot or similar ones. This can help us
in  defining cyclic structures like program above, without
using hacks like noop.

mdo extension (deprecated in 7.6 but undeprecate back in 7.8)
this is a syntactic sugar over mfix from MonadFix.

by wrinting
   mdo x <- create y
       y -> create x
       return (x,y)

we are creating structure:

   ( [   x   ]               [   y   ]
       ^  +-------create--------^  |
       +----------create-----------+